+++ src/main/java/com/example/hibernate/repository/CategoryRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     List<Category> findCategoryByTitleIgnoreCase(String title);
+
 }
