+++ src/main/java/com/example/hibernate/service/CategoryService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface CategoryService {
 
     List<Category> getAll();
-    Optional<Category> findById(int id);
+    Optional<Category> findById(long id);
     Optional<Category> findByTitle(String title);
     Category saveOrUpdate(Category category);
 }
